requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
lightgbm
matplotlib
seaborn
scipy
joblib
python-dotenv
openai>=1.0.0
shap
nbformat
torch
torchvision
Pillow
tqdm
ultralytics>=8.2.50
opencv-python
Pillow
prophet
pmdarima



